Enable persistence of search parameters [Fixes #9]
The biggest change here is the App component, which serves as the single entry point to the app (all remaining used endpoints besides index are now JSON API ones) and controls flow between our three main inter-related components (ModelSearcher, ModelTable, ModelDashboard). The basic logic is this:

App has a model id in local state. If it exists, the model dashboard is rendered with that model id. If not, the model searcher/table are rendered. Clicking on a model id in the table sets this model id state, and the back button removes it.

- Add App component which holds state and controls flow between ModelSearcher, ModelTable, and ModelDashboard
- Introduce ramda for easier transformations of data
- Introduce uniqueId utility to create unique keys for dynamically-created child nodes when there is no suitable domain-relevant id
- Modify MetricList and MetricSelector to control form through React state instead of jQuery DOM transformations
- Introduce ReactTable module and modify ModelTable to use it. Needed for greater control over cell rendering so React events can be triggered from it instead of just rendering HTML links
- Add CSS to control ReactTable pagination styles
- Remove now-unnecessary model prediction Flask route and template
- Remove jsx-handler-names eslint rule